Move GeoUtils to new geo module
