+++ .circleci/config.yml
@@ -93,6 +93,9 @@ jobs:
       - run:
           name: Run location unit tests
           command: ./gradlew -PdisablePreDex --no-daemon --max-workers=2 location:testDebug
+      - run:
+          name: Run geo unit tests
+          command: ./gradlew -PdisablePreDex --no-daemon --max-workers=2 geo:testDebug
 
       - store_artifacts:
           path: collect_app/build/reports
