+++ build.gradle
@@ -18,7 +18,7 @@ buildscript {
         classpath 'com.google.gms:google-services:4.3.5'
         classpath 'org.jacoco:org.jacoco.core:0.8.7'
         classpath 'com.google.firebase:firebase-crashlytics-gradle:2.7.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.kotlin}"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.21"
         classpath "org.jlleitschuh.gradle:ktlint-gradle:10.1.0"
         classpath "com.github.ben-manes:gradle-versions-plugin:0.39.0"
     }
