+++ collect_app/src/main/java/org/odk/collect/android/activities/FormMapActivity.java
@@ -39,7 +39,7 @@ import org.odk.collect.android.external.FormsContract;
 import org.odk.collect.android.external.InstanceProvider;
 import org.odk.collect.android.external.InstancesContract;
 import org.odk.collect.android.geo.MapFragment;
-import org.odk.collect.android.geo.MapPoint;
+import org.odk.collect.geo.MapPoint;
 import org.odk.collect.android.geo.MapProvider;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.preferences.keys.ProtectedProjectKeys;
