+++ collect_app/src/main/java/org/odk/collect/android/geo/MapFragment.java
@@ -21,6 +21,8 @@ import androidx.annotation.Nullable;
 import androidx.annotation.StringDef;
 import androidx.fragment.app.FragmentActivity;
 
+import org.odk.collect.geo.MapPoint;
+
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.util.List;
