+++ collect_app/src/test/java/org/odk/collect/android/geo/TestMapFragment.java
@@ -6,6 +6,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.FragmentActivity;
 
+import org.odk.collect.geo.MapPoint;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
