+++ geo/build.gradle
@@ -1,0 +1,44 @@
+import dependencies.Dependencies
+import dependencies.Versions
+
+plugins {
+    id 'com.android.library'
+    id 'kotlin-android'
+}
+
+apply from: '../config/quality.gradle'
+
+android {
+    compileSdkVersion Versions.android_compile_sdk
+
+    defaultConfig {
+        minSdkVersion Versions.android_min_sdk
+        targetSdkVersion Versions.android_target_sdk
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+        }
+    }
+
+    compileOptions {
+        coreLibraryDesugaringEnabled true
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+}
+
+dependencies {
+    coreLibraryDesugaring Dependencies.desugar
+
+    implementation Dependencies.kotlin_stdlib
+    implementation project(path: ':shared')
+}
