+++ settings.gradle
@@ -14,3 +14,4 @@ include ':nbistubs'
 include ':material'
 include ':fragmentstest'
 include ':location'
+include ':geo'
